+++ hexagon/core/src/test/kotlin/dev/dmigrate/core/diff/EnumCheckProjectionTest.kt
@@ -1,4 +1,4 @@
-package dev.dmigrate.core.diff.migration
+package dev.dmigrate.core.diff
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe